refactor: move trade-executor view models into models dir
