+++ src/lib/trade-executor/models/position-info.test.ts
