+++ src/lib/trade-executor/models/position-info.ts
