+++ src/lib/trade-executor/models/position-tooltips.ts
