+++ src/lib/trade-executor/models/trade-info.test.ts
