+++ src/lib/trade-executor/models/trade-info.ts
