+++ src/lib/trade-executor/models/trading-pair-info.ts
