+++ src/routes/strategies/[strategy]/[status=positionStatus]-positions/PositionTable.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { PositionStatus } from 'trade-executor/schemas/position';
-	import type { TradingPositionInfo } from 'trade-executor/state/position-info';
+	import type { TradingPositionInfo } from 'trade-executor/models/position-info';
 	import { writable } from 'svelte/store';
 	import { createTable, createRender } from 'svelte-headless-table';
 	import { addSortBy, addTableFilter, addColumnOrder, addPagination } from 'svelte-headless-table/plugins';
