+++ src/routes/strategies/[strategy]/[status=positionStatus]-positions/[position=integer]/+layout.ts
@@ -1,5 +1,5 @@
 import { error } from '@sveltejs/kit';
-import { getTradingPositionInfo } from 'trade-executor/state/position-info';
+import { getTradingPositionInfo } from 'trade-executor/models/position-info';
 
 export async function load({ params, parent }) {
 	// status can be `open`, `closed` or `frozen` (see params/positionStatus.ts)
