+++ src/routes/strategies/[strategy]/[status=positionStatus]-positions/[position=integer]/PositionProfitability.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-	import type { TradingPositionInfo } from 'trade-executor/state/position-info';
-	import { positionTooltips } from 'trade-executor/state/position-tooltips';
+	import type { TradingPositionInfo } from 'trade-executor/models/position-info';
+	import { positionTooltips } from 'trade-executor/models/position-tooltips';
 	import { DataBadge, Tooltip, UpDownIndicator } from '$lib/components';
 	import { formatProfitability } from 'trade-executor/helpers/formatters';
 	import { determineProfitability } from 'trade-executor/helpers/profit';
